+++ NEWS.md
@@ -1,8 +1,9 @@
 ## ggimage 0.1.8
 
-+ support `svg` file (2018-12-03, Mon)
++ support `pdf` file (2018-12-03, Mon)
++ support `svg` file 
   - <https://github.com/GuangchuangYu/ggimage/issues/11>
-  
+
 ## ggimage 0.1.7
 
 + `geom_worldcup2018` layer (2018-06-19, Tue)